requirements.txt
streamlit
python-docx
pandas
openpyxl
PyYAML
openai
